Refactor: separate inteface from the implementations
Just a tiny cleanup - no functional changes:
- keep the "shared/abstract" (config, interface) in ratelimit.go
- move the atomic implementation specific bits into "limiter_atomic"
- move "mutexbased" into "limiter_mutexbased" so that the naming
  is more consistent

Since I'm here I'm also updating the licenses.

The goal is to make reasoning/futher modifications easier, which
changes touching specific implementations only touching specific
implementation files.
